+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added option to export meta data for BLP files (disabled by default).
 - Added option to export meta data for WMO files (disabled by default).
 - Added 'Use Alpha' option to the Blender add-on. If unchecked, alpha channel is disabled and unlinked in the shader.
+- Renamed import menu option in Blender add-on from 'WoW OBJ (.obj) (experimental)' to 'WoW M2/WMO/ADT (.obj)'.
 - Layer meta data for ADT alpha maps is now exported in JSON format.
 - Textures for ADT alpha maps are now exported to their own relative path if 'Enabled Shared Textures' is enabled.
 - Textures for ADT alpha maps now have resolved names from the listfile, rather than fileDataIDs.